Перенос 3х пунктов меню
